@@ -1,6 +1,6 @@
 var fs = require("fs");
 var browserify = require("browserify");
-browserify("babel/master.jsx")
+browserify("babel/app.js")
   .transform("babelify", {presets: ["es2015", "react"]})
   .bundle()
   .pipe(fs.createWriteStream("public/js/bundle.js"));
